fix(smrt-svelte): make package browser-compatible
Move status enums (UserStatus, TenantStatus, etc.) from smrt-users to
smrt-types to enable browser-safe imports. This prevents browser builds
from pulling in Node.js-only dependencies through the transitive
dependency chain: smrt-svelte -> smrt-users -> smrt-core -> @happyvertical/files.

Changes:
- Add user status enums to smrt-types (browser-safe, no deps)
- Update smrt-users to re-export enums from smrt-types
- Update smrt-svelte to import UserStatus from smrt-types
- Move smrt-agents, smrt-profiles, smrt-users to optional peer deps in smrt-svelte
  (they're only used for type imports, not runtime code)

This fixes build failures in SvelteKit static sites that import from
@happyvertical/smrt-svelte when bundling for the browser.

Closes #681
